fix: extract climb_rate from OGN APRS packets
The climb_fpm field in the fixes table was not being populated because
the climb_rate data from OGN APRS packets was not being extracted.

Changes:
- Extract pos_packet.comment.climb_rate and cast to i32 for climb_fpm
- This will populate the Climb column on the flight details page fixes table

The OGN parser provides climb_rate as Option<i16> in fpm (feet per minute),
which is now properly extracted and stored in the database.