fix: replace zsh-specific >! with bash-compatible > in soar-deploy
The soar-deploy script uses #!/bin/bash but was using >! syntax for
file redirection, which is zsh-specific (for overriding noclobber).

In bash, >! is not recognized as a redirect operator. This caused bash
to parse the redirect target as an additional input file argument to sed,
resulting in errors like:
  sed: can't read /etc/grafana/provisioning/alerting/contact-points.yml:
  No such file or directory

Changed both instances of >! to > (standard bash redirect), which works
correctly since bash doesn't have noclobber enabled by default.

This fixes staging deployment failures when installing Grafana alerting
configuration files.